getData/requirements.txt
requests-html
emoji
nltk
sklearn
requests-html
emoji
tensorflow
numpy
scipy
opencv-python
pillow
matplotlib
h5py
keras
https://github.com/OlafenwaMoses/ImageAI/releases/download/2.0.2/imageai-2.0.2-py3-none-any.whl 
selenium
pandas